Split Main and Pause Menus Into Their Own Modules
